graphrag-system/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Embeddings & ML
sentence-transformers>=2.2.0
torch>=2.0.0

# Graph
networkx>=3.0
python-igraph>=0.10.0
leidenalg>=0.10.0

# Vector storage (optional, for production)
# faiss-cpu>=1.7.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# LLM
requests>=2.31.0

# Data formats
pyarrow>=12.0.0

# Utilities
tqdm>=4.65.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
